Refactor models into separate modules for better organization
- Split `models.py` into multiple files: `configuration.py`, `claude.py`, `github.py`, `workflow.py`
- Improve maintainability by separating concerns within the `imbi_automations` package
- No functional changes; only structural refactor for clarity and modularity
